requirements/base.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
channels==4.2.0
channels_redis==4.2.1
daphne==4.1.2
celery==5.4.0
django-celery-beat==2.7.0
django-cors-headers==4.4.0
psycopg2==2.9.10
redis==5.2.1
python-dateutil==2.9.0.post0
sqlparse==0.5.1
tzdata==2025.1
amqp==5.3.1
billiard==4.2.1
kombu==5.5.0
Twisted==24.11.0
txaio==23.1.1
autobahn==24.4.2
hyperlink==21.0.0
constantly==23.10.4
cffi==1.17.1
cryptography==44.0.2
pycparser==2.22
pyOpenSSL==25.0.0
service-identity==24.2.0
attrs==25.3.0
vine==5.1.0
six==1.17.0
idna==3.10
incremental==24.7.2
packaging==24.2
pydantic==2.11.9
pydantic_core==2.33.2
pika
